Forward compatibility with Rails 4.1+
Concerning was merged with Rails 4.1. To make life easy on libraries
that support multiple versions of Rails, this lib will step aside if it
sees that the Active Support implementation is available.

That way your lib can depend on this gem and `require 'concerning'`
without worrying about versioning issues or implementation collision.